+++ p_maputl.c
@@ -373,8 +373,8 @@ boolean ThingIsOnLine(mobj_t *t, line_t *l)
 
 boolean P_BlockLinesIterator(int x, int y, boolean func(line_t*))
 {
-  int        offset;
-  const long *list;   // killough 3/1/98: for removal of blockmap limit
+  int             offset;
+  const long long *list;   // killough 3/1/98: for removal of blockmap limit
 
   if (x<0 || y<0 || x>=bmapwidth || y>=bmapheight)
     return true;
