+++ win_opendir.c
@@ -273,7 +273,7 @@ void rewinddir(DIR * dirp)
 // Returns the "position" in the "directory stream" which can be used with
 // seekdir to go back to an old entry. We simply return the value in stat.
 //
-long telldir(DIR *dirp)
+long long telldir(DIR *dirp)
 {
    errno = 0;
    
@@ -294,7 +294,7 @@ long telldir(DIR *dirp)
 // have changed while we weren't looking. But that is probably the case with
 // any such system.
 //
-void seekdir(DIR *dirp, long lPos)
+void seekdir(DIR *dirp, long long lPos)
 {
    errno = 0;
